requirements.txt
flask>=2.3.0
langchain-google-genai>=0.2.0
langgraph>=0.4.0
langgraph-checkpoint-postgres>=2.0.23
langchain-core>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.0

